[EI-253] [api] rename: test data 명 변경
