+++ apps/api/src/numerical-guidance/test/E2E test/indicator/indicator.e2e.spec.ts
@@ -22,11 +22,11 @@ import * as fs from 'fs';
 import { GetIndicatorListQueryHandler } from '../../../application/query/indicator/get-indicator-list/get-indicator-list.query.handler';
 import { SearchTwelveIndicatorQueryHandler } from '../../../application/query/indicator/search-twelve-indicator/search-twelve-indicator.query.handler';
 import { IndicatorTwelveAdapter } from '../../../infrastructure/adapter/twelve/indicator.twelve.adapter';
-import { TwelveApiUtil } from '../../../infrastructure/adapter/twelve/util/twelve-api.util';
+import { TwelveApiManager } from '../../../infrastructure/adapter/twelve/util/twelve-api.manager';
 import { AdjustIndicatorValue } from '../../../util/adjust-indicator-value';
 import { SearchIndicatorQueryHandler } from 'src/numerical-guidance/application/query/indicator/search-indicator/search-indicator.query.handler';
 import { EconomyEntity } from '../../../infrastructure/adapter/persistence/indicator/entity/economy.entity';
-import { FredApiUtil } from '../../../infrastructure/adapter/fred/util/fred-api.util';
+import { FredApiManager } from '../../../infrastructure/adapter/fred/util/fred-api.manager';
 
 const filePath = './src/numerical-guidance/test/data/indicator-list-stocks.json';
 const data = fs.readFileSync(filePath, 'utf8');
@@ -125,8 +125,8 @@ describe('Indicator E2E Test', () => {
           GetIndicatorListQueryHandler,
           SearchTwelveIndicatorQueryHandler,
           SearchIndicatorQueryHandler,
-          TwelveApiUtil,
-          FredApiUtil,
+          TwelveApiManager,
+          FredApiManager,
           {
             provide: 'LoadIndicatorPort',
             useClass: IndicatorPersistentAdapter,
