+++ apps/api/src/numerical-guidance/test/data/live-indicator-test-data-per-day.json
