+++ apps/api/src/numerical-guidance/test/data/live-indicator-test-data-per-month.json
